+++ internal/utils/utils.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"path"
 	"slices"
 	"strconv"
 	"strings"
@@ -61,32 +60,6 @@ func IsPersistentVolumeClaimConditionPresentAndEqual(obj *corev1.PersistentVolum
 	return false
 }
 
-// HasSchedulingGate returns whether the given Pod already has a scheduling gate
-// with the given name in its spec.
-func HasSchedulingGate(pod *corev1.Pod, name string) bool {
-	return slices.ContainsFunc(pod.Spec.SchedulingGates, func(gate corev1.PodSchedulingGate) bool {
-		return gate.Name == name
-	})
-}
-
-// AddSchedulingGate adds a scheduling gate with the given name to the Pod's spec,
-// if it is not already present.
-func AddSchedulingGate(pod *corev1.Pod, name string) {
-	if HasSchedulingGate(pod, name) {
-		return
-	}
-
-	pod.Spec.SchedulingGates = append(pod.Spec.SchedulingGates, corev1.PodSchedulingGate{Name: name})
-}
-
-// RemoveSchedulingGate removes the scheduling gate with the given name from the
-// Pod's spec if it is present.
-func RemoveSchedulingGate(pod *corev1.Pod, name string) {
-	pod.Spec.SchedulingGates = slices.DeleteFunc(pod.Spec.SchedulingGates, func(gate corev1.PodSchedulingGate) bool {
-		return gate.Name == name
-	})
-}
-
 // FindOwningPVCAAndPolicy returns the PersistentVolumeClaimAutoscaler that manages the
 // given PersistentVolumeClaim and whose spec.autoscalerName matches the given
 // autoscalerName, together with the VolumePolicy that applies to the PVC, or
@@ -97,9 +70,7 @@ func FindOwningPVCAAndPolicy(ctx context.Context, c client.Client, autoscalerNam
 		return nil, nil, fmt.Errorf("failed to list PersistentVolumeClaimAutoscalers: %w", err)
 	}
 
-	for i := range pvcaList.Items {
-		pvca := &pvcaList.Items[i]
-
+	for _, pvca := range pvcaList.Items {
 		if slices.ContainsFunc(pvca.Status.VolumeRecommendations, func(vr v1alpha1.VolumeRecommendation) bool {
 			return vr.Name == pvc.Name
 		}) {
@@ -108,38 +79,9 @@ func FindOwningPVCAAndPolicy(ctx context.Context, c client.Client, autoscalerNam
 				return nil, nil, err
 			}
 
-			return pvca, policy, nil
+			return &pvca, policy, nil
 		}
 	}
 
 	return nil, nil, nil
-}
-
-// GetVolumePolicy returns the VolumePolicy for the given PersistentVolumeClaim
-// name. Policies are evaluated in list order and the first policy whose match
-// glob matches the name is returned. It returns nil if no policy matches.
-func GetVolumePolicy(pvcName string, volumePolicies []v1alpha1.VolumePolicy) (*v1alpha1.VolumePolicy, error) {
-	for i := range volumePolicies {
-		matched, err := path.Match(volumePolicies[i].Match.Name, pvcName)
-		if err != nil {
-			return nil, fmt.Errorf("invalid volume policy name %q: %w", volumePolicies[i].Match.Name, err)
-		}
-		if matched {
-			return &volumePolicies[i], nil
-		}
-	}
-
-	return nil, nil
-}
-
-// IsOfflineResizeEnabled reports whether the given PersistentVolumeClaimAutoscaler
-// and the VolumePolicy that applies to a PVC it manages opt into offline-resize
-// recovery. This requires that both both the policy's scaleUp.resizeStrategy is PreferInPlace,
-// and the PVCA targets a workload controller rather than a PersistentVolumeClaim directly.
-func IsOfflineResizeEnabled(pvca *v1alpha1.PersistentVolumeClaimAutoscaler, policy *v1alpha1.VolumePolicy) bool {
-	return pvca != nil &&
-		pvca.Spec.TargetRef.Kind != "PersistentVolumeClaim" &&
-		policy != nil &&
-		policy.ScaleUp != nil &&
-		policy.ScaleUp.ResizeStrategy == v1alpha1.PreferInPlaceVolumeResizeStrategy
 }
