+++ internal/utils/utils_test.go
@@ -87,199 +87,6 @@ var _ = Describe("Utils", func() {
 		})
 	})
 
-	Context("scheduling gates", func() {
-		const gate = "pvc.autoscaling.gardener.cloud/offline-resize"
-
-		var newPod = func(gates ...string) *corev1.Pod {
-			pod := &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "sample-pod", Namespace: "default"},
-			}
-			for _, gate := range gates {
-				pod.Spec.SchedulingGates = append(pod.Spec.SchedulingGates, corev1.PodSchedulingGate{Name: gate})
-			}
-
-			return pod
-		}
-
-		Describe("#HasSchedulingGate", func() {
-			It("should return true when the gate is present", func() {
-				Expect(utils.HasSchedulingGate(newPod("other", gate), gate)).To(BeTrue())
-			})
-
-			It("should return false when the gate is absent", func() {
-				Expect(utils.HasSchedulingGate(newPod("other"), gate)).To(BeFalse())
-			})
-		})
-
-		Describe("#AddSchedulingGate", func() {
-			It("should add the gate and report modification", func() {
-				pod := newPod("other")
-				utils.AddSchedulingGate(pod, gate)
-
-				Expect(pod.Spec.SchedulingGates).To(HaveLen(2))
-				Expect(pod.Spec.SchedulingGates).To(ContainElement(HaveField("Name", gate)))
-			})
-
-			It("should be idempotent when the gate is already present", func() {
-				pod := newPod(gate)
-				utils.AddSchedulingGate(pod, gate)
-
-				Expect(pod.Spec.SchedulingGates).To(HaveLen(1))
-			})
-		})
-
-		Describe("#RemoveSchedulingGate", func() {
-			It("should remove the gate and report modification", func() {
-				pod := newPod("other", gate)
-				utils.RemoveSchedulingGate(pod, gate)
-
-				Expect(pod.Spec.SchedulingGates).To(HaveLen(1))
-				Expect(pod.Spec.SchedulingGates).To(ConsistOf(HaveField("Name", "other")))
-			})
-
-			It("should be a no-op when the gate is absent", func() {
-				pod := newPod("other")
-				utils.RemoveSchedulingGate(pod, gate)
-
-				Expect(pod.Spec.SchedulingGates).To(HaveLen(1))
-				Expect(pod.Spec.SchedulingGates).To(ConsistOf(HaveField("Name", "other")))
-			})
-		})
-	})
-
-	Describe("#GetVolumePolicy", func() {
-		It("should return error on invalid glob pattern", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "[",
-					},
-					MaxCapacity: resource.MustParse("10Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).To(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring("invalid volume policy name \"[\""))
-			Expect(policy).To(BeNil())
-		})
-
-		It("should return nil when no policy matches", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "other-pvc",
-					},
-					MaxCapacity: resource.MustParse("10Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).To(BeNil())
-		})
-
-		It("should match exact name", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "data-pvc",
-					},
-					MaxCapacity: resource.MustParse("10Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).NotTo(BeNil())
-			Expect(policy.Match.Name).To(Equal("data-pvc"))
-		})
-
-		It("should match glob pattern", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "*-logs",
-					},
-					MaxCapacity: resource.MustParse("15Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("app-logs", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).NotTo(BeNil())
-			Expect(policy.Match.Name).To(Equal("*-logs"))
-		})
-
-		It("should match default policy", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "*",
-					},
-					MaxCapacity: resource.MustParse("5Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).NotTo(BeNil())
-			Expect(policy.Match.Name).To(Equal("*"))
-		})
-
-		It("should fall back to default policy when no other policy matches", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "other-pvc",
-					},
-					MaxCapacity: resource.MustParse("20Gi"),
-				},
-				{
-					Match: v1alpha1.Match{
-						Name: "*",
-					},
-					MaxCapacity: resource.MustParse("5Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).NotTo(BeNil())
-			Expect(policy.Match.Name).To(Equal("*"))
-			Expect(policy.MaxCapacity).To(Equal(resource.MustParse("5Gi")))
-		})
-
-		It("should return first seen matching policy", func() {
-			volumePolicies := []v1alpha1.VolumePolicy{
-				{
-					Match: v1alpha1.Match{
-						Name: "data-*",
-					},
-					MaxCapacity: resource.MustParse("10Gi"),
-				},
-				{
-					Match: v1alpha1.Match{
-						Name: "data-pvc",
-					},
-					MaxCapacity: resource.MustParse("20Gi"),
-				},
-				{
-					Match: v1alpha1.Match{
-						Name: "*",
-					},
-					MaxCapacity: resource.MustParse("5Gi"),
-				},
-			}
-
-			policy, err := utils.GetVolumePolicy("data-pvc", volumePolicies)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(policy).NotTo(BeNil())
-			Expect(policy.Match.Name).To(Equal("data-*"))
-			Expect(policy.MaxCapacity).To(Equal(resource.MustParse("10Gi")))
-		})
-	})
-
 	Describe("#FindOwningPVCAAndPolicy", func() {
 		const (
 			namespace      = "default"
@@ -360,53 +167,6 @@ var _ = Describe("Utils", func() {
 				p.Status.VolumeRecommendations = []v1alpha1.VolumeRecommendation{{Name: "other-pvc"}}
 			}, true, false),
 			Entry("when the PVCA is in a different namespace", func(p *v1alpha1.PersistentVolumeClaimAutoscaler) { p.Namespace = "other-namespace" }, true, false),
-		)
-	})
-
-	Context("#IsOfflineResizeEnabled", func() {
-		var (
-			workloadPVCA *v1alpha1.PersistentVolumeClaimAutoscaler
-			preferPolicy *v1alpha1.VolumePolicy
-		)
-
-		BeforeEach(func() {
-			workloadPVCA = &v1alpha1.PersistentVolumeClaimAutoscaler{
-				Spec: v1alpha1.PersistentVolumeClaimAutoscalerSpec{
-					TargetRef: autoscalingv1.CrossVersionObjectReference{Kind: "StatefulSet"},
-				},
-			}
-			preferPolicy = &v1alpha1.VolumePolicy{ScaleUp: &v1alpha1.ScalingRules{ResizeStrategy: v1alpha1.PreferInPlaceVolumeResizeStrategy}}
-		})
-
-		DescribeTable("should report whether offline-resize recovery is enabled",
-			func(mutate func(*v1alpha1.PersistentVolumeClaimAutoscaler, *v1alpha1.VolumePolicy), want bool) {
-				mutate(workloadPVCA, preferPolicy)
-				Expect(utils.IsOfflineResizeEnabled(workloadPVCA, preferPolicy)).To(Equal(want))
-			},
-			Entry("when the PVCA targets a workload and the strategy is PreferInPlace",
-				func(*v1alpha1.PersistentVolumeClaimAutoscaler, *v1alpha1.VolumePolicy) {}, true),
-			Entry("when the PVCA is nil",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, _ *v1alpha1.VolumePolicy) { workloadPVCA = nil }, false),
-			Entry("when the PVCA targets a PersistentVolumeClaim directly",
-				func(p *v1alpha1.PersistentVolumeClaimAutoscaler, _ *v1alpha1.VolumePolicy) {
-					p.Spec.TargetRef.Kind = "PersistentVolumeClaim"
-				}, false),
-			Entry("when the policy is nil",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, _ *v1alpha1.VolumePolicy) { preferPolicy = nil }, false),
-			Entry("when scaleUp is nil",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, p *v1alpha1.VolumePolicy) { p.ScaleUp = nil }, false),
-			Entry("when the strategy is unset",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, p *v1alpha1.VolumePolicy) {
-					p.ScaleUp.ResizeStrategy = ""
-				}, false),
-			Entry("when the strategy is InPlace",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, p *v1alpha1.VolumePolicy) {
-					p.ScaleUp.ResizeStrategy = v1alpha1.InPlaceVolumeResizeStrategy
-				}, false),
-			Entry("when the strategy is Off",
-				func(_ *v1alpha1.PersistentVolumeClaimAutoscaler, p *v1alpha1.VolumePolicy) {
-					p.ScaleUp.ResizeStrategy = v1alpha1.OffVolumeResizeStrategy
-				}, false),
 		)
 	})
 })
